ai-ddr-generator/test.py
from groq import Groq
import os
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("GROQ_API_KEY")
client = Groq(api_key=api_key)

try:
    response = client.chat.completions.create(
        model="llama-3.3-70b-versatile",
        messages=[{"role": "user", "content": "hello"}]
    )
    print(response.choices[0].message.content)
except Exception as e:
    print(f"API ERROR: {e}")
