ai-ddr-generator/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.2.2
PyMuPDF==1.25.2
streamlit==1.42.0
groq==0.11.0
